# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CinemaView:
    properties:
      actors:
        items:
          type: string
        type: array
      age_rating:
        type: integer
      description:
        type: string
      duration_minutes:
        type: integer
      genres:
        items:
          type: string
        type: array
      id:
        type: string
      name:
        type: string
      poster_url:
        type: string
      rating:
        type: number
      user_rating:
        type: integer
      year:
        type: integer
    type: object
  dto.CouchView:
    properties:
      author:
        type: string
      id:
        type: string
      name:
        type: string
      users:
        items:
          type: string
        type: array
    type: object
  dto.CreateCouch:
    properties:
      name:
        type: string
      users:
        items:
          type: string
        type: array
    type: object
  dto.HttpErr:
    properties:
      description:
        example: verbose error description
        type: string
      err:
        example: 'Short error message : ''Not Found'' | ''Internal Server Error''
          | etc'
        type: string
    required:
    - description
    - err
    type: object
...
  dto.UpdateCouch:
    properties:
      name:
        type: string
      users:
        items:
          type: string
        type: array
    type: object
...
host: https://prod-team-32-n26k57br.REDACTED
info:
  contact: {}
  title: MetaCinema
  version: "1.0"
paths:
...
  /couches/:
    get:
      description: Получает все диваны
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CouchView'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: get couches
      tags:
      - couches
    post:
      consumes:
      - application/json
      description: Создает новый диван
      parameters:
      - description: Request Body
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCouch'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CouchView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: create new couch
      tags:
      - couches
  /couches/{id}:
    get:
      description: Получает диван по его ID
      parameters:
      - description: Couch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CouchView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: get couch
      tags:
      - couches
    put:
      consumes:
      - application/json
      description: Обновляет диван
      parameters:
      - description: Couch ID
        in: path
        name: id
        required: true
        type: string
      - description: Request Body
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCouch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CouchView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: update couch
      tags:
      - couches
  /couches/{id}/feed:
    get:
      description: Показывает рекомендации в диване
      parameters:
      - description: Couch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CinemaView'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: get recommended films in couch
      tags:
      - couches
  /couches/{id}/films/{filmId}/dislike:
    delete:
      description: Удаляет дизлайк фильма в диване
      parameters:
      - description: Couch ID
        in: path
        name: id
        required: true
        type: string
      - description: Film ID
        in: path
        name: filmId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: delete dislike of the film in couch
      tags:
      - couches
    post:
      description: Дизлайкает фильм в диване
      parameters:
      - description: Couch ID
        in: path
        name: id
        required: true
        type: string
      - description: Film ID
        in: path
        name: filmId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: dislike film in couch
      tags:
      - couches
  /couches/{id}/films/{filmId}/like:
    delete:
      description: Удаляет лайк фильма в диване
      parameters:
      - description: Couch ID
        in: path
        name: id
        required: true
        type: string
      - description: Film ID
        in: path
        name: filmId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: delete like of the film in couch
      tags:
      - couches
    post:
      description: Лайкает фильм в диване
      parameters:
      - description: Couch ID
        in: path
        name: id
        required: true
        type: string
      - description: Film ID
        in: path
        name: filmId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: like film in couch
      tags:
      - couches
  /couches/{id}/plans:
    get:
      description: Получает любимые фильмы дивана
      parameters:
      - description: Couch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CouchView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: get couch likes
      tags:
      - couches
  /couches/{id}/views/bulk:
    post:
      parameters:
      - description: Couch ID
        in: path
        name: id
        required: true
        type: string
      - description: Request Body
        in: body
        name: RequestBody
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: mark film as seen in couch
      tags:
      - couches
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"